Splits the project into `core` and `cli` modules
- Uses `lerna` to manage modules
- Adds README.md for each module
- Updates version to 0.4.0
