--- ROADMAP.md
@@ -1,9 +1,0 @@
-#Roadmap
-
-Not in any specific order:
-
-1. Updatable builds: `llama-pack update` updates to latest version of the template project.
-2. Installs dependencies to build the Android Project automatically (JDK, ADT and Gradle)
-5. Running Lighthouse
-6. [Adaptive Icons](https://developer.android.com/guide/practices/ui_guidelines/icon_design_adaptive)
-7. Generate `/.well-know/assetlinks.json`. (depends on retrieving SHA-256 from key)
