+++ package.json
@@ -1,15 +1,13 @@
 {
-  "name": "llama-pack",
-  "version": "0.3.0",
-  "description": "CLI tool to Generate TWA projects from a Web Manifest",
-  "bin": "bin/llama-pack.js",
+  "name": "root",
+  "private": true,
+  "version": "0.4.1",
+  "description": "Generate TWA projects from a Web Manifest",
   "scripts": {
-    "postinstall": "tsc",
-    "build": "tsc",
-    "lint": "eslint \"src/**/*.{js,ts}\"",
-    "test": "tsc && jasmine --config=jasmine.json"
+    "build": "lerna run build",
+    "lint": "lerna run lint",
+    "test": "lerna run test"
   },
-  "main": "dist/index.js",
   "keywords": [
     "twa",
     "trusted-web-activities",
@@ -24,33 +22,17 @@
   "author": "",
   "license": "Apache-2.0",
   "dependencies": {
-    "@types/color": "^3.0.0",
-    "@types/extract-zip": "^1.6.2",
-    "@types/inquirer": "^6.5.0",
-    "@types/jasmine": "^3.5.0",
-    "@types/lodash.template": "^4.4.6",
-    "@types/minimist": "^1.2.0",
-    "@types/node": "^12.12.14",
-    "@types/node-fetch": "^2.5.4",
-    "@types/tar": "^4.0.3",
-    "@types/valid-url": "^1.0.2",
-    "color": "^3.1.2",
-    "extract-zip": "^1.6.7",
-    "inquirer": "^7.0.4",
-    "jimp": "^0.9.3",
-    "lodash.template": "^4.5.0",
-    "minimist": "^1.2.0",
-    "node-fetch": "^2.6.0",
-    "npm": "^6.13.7",
-    "tar": "^5.0.5",
-    "typescript": "^3.7.2",
-    "valid-url": "^1.0.9"
+    "@llama-pack/core": "file:packages/core",
+    "@llama-pack/cli": "file:packages/cli"
   },
   "devDependencies": {
     "@typescript-eslint/eslint-plugin": "^2.9.0",
     "@typescript-eslint/parser": "^2.9.0",
     "eslint": "^6.6.0",
     "eslint-config-google": "^0.14.0",
-    "jasmine": "^3.5.0"
+    "jasmine": "^3.5.0",
+    "lerna": "^3.20.2",
+    "lerna-changelog": "^1.0.0",
+    "typescript": "^3.7.2"
   }
 }
