+++ packages/cli/src/index.ts
@@ -1,0 +1,29 @@
+/*
+ * Copyright 2019 Google Inc. All Rights Reserved.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+import {Cli} from './lib/Cli';
+import {Log} from '@llama-pack/core';
+
+module.exports = (): void => {
+  const cli = new Cli();
+  const log = new Log('cli');
+  const args = process.argv.slice(2);
+  cli.run(args)
+      .catch((err: Error) => {
+        log.error(err.message);
+        process.exit(1);
+      });
+};
