+++ packages/core/jasmine.json
@@ -1,0 +1,11 @@
+{
+  "spec_dir": "dist/spec",
+  "spec_files": [
+    "**/*[sS]pec.js"
+  ],
+  "helpers": [
+    "helpers/**/*.js"
+  ],
+  "stopSpecOnExpectationFailure": false,
+  "random": true
+}
