+++ packages/core/src/lib/Config.ts
