+++ packages/core/src/lib/types/WebManifest.ts
