+++ packages/core/src/spec/lib/ConfigSpec.ts
