+++ packages/core/src/spec/lib/TwaManifestSpec.ts
