+++ packages/core/src/spec/lib/androidSdk/AndroidSdkToolsSpec.ts
