+++ packages/core/template_project/app/.gitignore
