+++ packages/core/template_project/app/build.gradle
