+++ packages/core/template_project/app/src/main/AndroidManifest.xml
