+++ packages/core/template_project/app/src/main/res/mipmap-anydpi-v26/ic_launcher.xml
