+++ packages/core/template_project/app/src/main/res/values/styles.xml
