+++ packages/core/template_project/app/src/main/res/xml/filepaths.xml
