+++ packages/core/template_project/build.gradle
