+++ packages/core/template_project/enable-debug.sh
