+++ packages/core/template_project/gradle.properties
