+++ packages/core/template_project/gradlew.bat
