+++ packages/core/template_project/settings.gradle
