+++ packages/core/tsconfig.json
@@ -1,0 +1,12 @@
+{
+  "compilerOptions": {
+    "allowJs": true,
+    "target": "es5",
+    "module": "commonjs",
+    "declaration": true,
+    "outDir": "dist",
+    "strict": true
+  },
+  "include": ["src/**/*.ts"],  
+  "exclude": ["node_modules", "dist"]
+}
